+++ llx/tools/__init__.py
@@ -19,4 +19,4 @@ __all__ = [
     "HealthChecker"
 ]
 
-__version__ = "0.1.71"
+__version__ = "0.1.72"